breakup create tournament form file into smaller files
