+++ cspell.config.yaml
@@ -41,6 +41,7 @@ words:
   - adtest
   - alola
   - battlestadium
+  - checkin
   - cmdk
   - Embla
   - galar
@@ -70,6 +71,7 @@ words:
   - Sonner
   - sqlstring
   - teamlists
+  - teamsheet
   - Tera
   - thatguyinabeanie
   - typesafe
